+++ GoDiveMVP/Models/Certification+CardType.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+extension Certification {
+    var cardType: CertificationCardType {
+        get { CertificationCardType(rawValue: cardTypeRaw) ?? .certification }
+        set { cardTypeRaw = newValue.rawValue }
+    }
+}